Unificacion Microservicio de Pago Electronico. Primera version del derivador

+++ code/back/microservices/ms-pagoelectronico/src/main/java/com/topicosplmspagoelectronico/bean/dto/PaymentDTO.java
@@ -10,7 +10,7 @@ public class PaymentDTO {
     private String numeroTarjeta;
     private String FechaExpiracion;
     private String NumeroDeVerificacion;
-    private int cuotas;
-    private int total;
+    private Long cuotas;
+    private Double total;
 
 }
